+++ packages/core/src/index.ts
@@ -1,12 +1,14 @@
-import type { CeramicApi, Doctype } from '@ceramicnetwork/common'
-import type DocID from '@ceramicnetwork/docid'
+import type { CeramicApi } from '@ceramicnetwork/common'
+import type StreamID from '@ceramicnetwork/streamid'
+import { TileDocument } from '@ceramicnetwork/stream-tile'
+import { Caip10Link } from '@ceramicnetwork/stream-caip10-link'
 import { definitions, schemas } from '@ceramicstudio/idx-constants'
-import DataLoader from 'dataloader'
 
-import { DoctypeProxy } from './doctypes'
+import { DoctypeProxy, TileDoc, TileContent } from './doctypes'
 import type { Aliases, DefinitionWithID, Entry, Index, IndexKey } from './types'
-import { toDocIDString } from './utils'
+import { isCaip10, assertDid } from './utils'
 
+export { getLegacy3BoxProfileAsBasicProfile } from './3box'
 export * from './types'
 export * from './utils'
 
@@ -24,17 +26,12 @@ export class IDX {
   _aliases: Aliases
   _autopin: boolean
   _ceramic: CeramicApi
-  _docLoader: DataLoader<string, Doctype>
   _indexProxy: DoctypeProxy
 
   constructor({ aliases = {}, autopin, ceramic }: IDXOptions) {
     this._aliases = { ...definitions, ...aliases }
     this._autopin = autopin !== false
     this._ceramic = ceramic
-
-    this._docLoader = new DataLoader(async (ids: ReadonlyArray<string>) => {
-      return await Promise.all(ids.map(async (id) => await this._ceramic.loadDocument<Doctype>(id)))
-    })
     this._indexProxy = new DoctypeProxy(this._getOwnIDXDoc.bind(this))
   }
 
@@ -66,7 +63,11 @@ export class IDX {
     return await this._getRecord(key, did)
   }
 
-  async set(name: string, content: unknown, options?: CreateOptions): Promise<DocID> {
+  async set(
+    name: string,
+    content: Record<string, any>,
+    options?: CreateOptions
+  ): Promise<StreamID> {
     const key = this._toIndexKey(name)
     return await this._setRecord(key, content, options)
   }
@@ -75,19 +76,22 @@ export class IDX {
     name: string,
     content: T,
     options?: CreateOptions
-  ): Promise<DocID> {
+  ): Promise<StreamID> {
     const key = this._toIndexKey(name)
     const existing = await this._getRecord<T>(key)
     const newContent = existing ? { ...existing, ...content } : content
     return await this._setRecord(key, newContent, options)
   }
 
-  async setAll(contents: Record<string, unknown>, options?: CreateOptions): Promise<Index> {
+  async setAll(
+    contents: Record<string, Record<string, any>>,
+    options?: CreateOptions
+  ): Promise<Index> {
     const updates = Object.entries(contents).map(async ([name, content]) => {
       const key = this._toIndexKey(name)
       const [created, id] = await this._setRecordOnly(key, content, options)
       return [created, key, id]
-    }) as Array<Promise<[boolean, IndexKey, DocID]>>
+    }) as Array<Promise<[boolean, IndexKey, StreamID]>>
     const changes = await Promise.all(updates)
 
     const newReferences = changes.reduce((acc, [created, key, id]) => {
@@ -101,16 +105,23 @@ export class IDX {
     return newReferences
   }
 
-  async setDefaults(contents: Record<string, unknown>, options?: CreateOptions): Promise<Index> {
+  async setDefaults(
+    contents: Record<string, Record<string, any>>,
+    options?: CreateOptions
+  ): Promise<Index> {
     const index = (await this.getIndex()) ?? {}
 
     const updates = Object.entries(contents)
       .map(([name, content]) => [this._toIndexKey(name), content])
-      .filter(([key]) => index[key as IndexKey] == null)
+      // This filter returned the type (string | Record<string, any>)[][]
+      // we need to add a type guard to get the correct type here.
+      .filter(
+        (entry): entry is [string, Record<string, any>] => index[entry[0] as IndexKey] == null
+      )
       .map(async ([key, content]) => {
-        const definition = await this.getDefinition(key as IndexKey)
+        const definition = await this.getDefinition(key)
         const id = await this._createRecord(definition, content, options)
-        return { [key as IndexKey]: id.toUrl() }
+        return { [key]: id.toUrl() }
       }) as Array<Promise<Index>>
     const changes = await Promise.all(updates)
 
@@ -165,22 +176,20 @@ export class IDX {
     }
   }
 
-  async _createIDXDoc(did: string): Promise<Doctype> {
-    const doc = await this._ceramic.createDocument(
-      'tile',
-      {
-        deterministic: true,
-        metadata: { controllers: [did], family: 'IDX' },
-      },
+  async _createIDXDoc(did: string): Promise<TileDoc> {
+    if (isCaip10(did)) {
+      did = await this.caip10ToDid(did)
+    }
+    assertDid(did)
+    return await TileDocument.create<TileContent>(
+      this._ceramic,
+      null,
+      { deterministic: true, controllers: [did], family: 'IDX' },
       { anchor: false, publish: false }
     )
-    if (this._autopin) {
-      await this._ceramic.pin.add(doc.id)
-    }
-    return doc
   }
 
-  async _getIDXDoc(did: string): Promise<Doctype | null> {
+  async _getIDXDoc(did: string): Promise<TileDoc | null> {
     const doc = await this._createIDXDoc(did)
     if (doc.metadata.schema == null) {
       return null
@@ -191,11 +200,14 @@ export class IDX {
     return doc
   }
 
-  async _getOwnIDXDoc(): Promise<Doctype> {
+  async _getOwnIDXDoc(): Promise<TileDoc> {
     const doc = await this._createIDXDoc(this.id)
     if (doc.metadata.schema == null) {
       // Doc just got created, need to update it with schema
-      await doc.change({ metadata: { ...doc.metadata, schema: schemas.IdentityIndex } })
+      await doc.update(null, { ...doc.metadata, schema: schemas.IdentityIndex })
+      if (this._autopin) {
+        await this._ceramic.pin.add(doc.id)
+      }
     } else if (doc.metadata.schema !== schemas.IdentityIndex) {
       throw new Error('Invalid document: schema is not IdentityIndex')
     }
@@ -204,7 +216,7 @@ export class IDX {
 
   // Definition APIs
 
-  async getDefinition(idOrKey: DocID | IndexKey): Promise<DefinitionWithID> {
+  async getDefinition(idOrKey: StreamID | IndexKey): Promise<DefinitionWithID> {
     const doc = await this._loadDocument(idOrKey)
     if (doc.metadata.schema !== schemas.Definition) {
       throw new Error('Invalid document: schema is not Definition')
@@ -218,7 +230,7 @@ export class IDX {
     return await this._getReference(key, did)
   }
 
-  async getRecordDocument(key: IndexKey, did?: string): Promise<Doctype | null> {
+  async getRecordDocument(key: IndexKey, did?: string): Promise<TileDoc | null> {
     const id = await this.getRecordID(key, did)
     if (id == null) {
       return null
@@ -231,7 +243,11 @@ export class IDX {
     return doc ? (doc.content as T) : null
   }
 
-  async _setRecord(key: IndexKey, content: unknown, options?: CreateOptions): Promise<DocID> {
+  async _setRecord(
+    key: IndexKey,
+    content: Record<string, any>,
+    options?: CreateOptions
+  ): Promise<StreamID> {
     const [created, id] = await this._setRecordOnly(key, content, options)
     if (created) {
       await this._setReference(key, id)
@@ -241,9 +257,9 @@ export class IDX {
 
   async _setRecordOnly(
     key: IndexKey,
-    content: unknown,
+    content: Record<string, any>,
     { pin }: CreateOptions = {}
-  ): Promise<[boolean, DocID]> {
+  ): Promise<[boolean, StreamID]> {
     const existing = await this._getReference(key, this.id)
     if (existing == null) {
       const definition = await this.getDefinition(key)
@@ -251,29 +267,32 @@ export class IDX {
       return [true, ref]
     } else {
       const doc = await this._loadDocument(existing)
-      await doc.change({ content })
+      await doc.update(content)
       return [false, doc.id]
     }
   }
 
-  async _loadDocument(id: DocID | string): Promise<Doctype> {
-    return await this._docLoader.load(toDocIDString(id))
+  _loadDocument(id: StreamID | string): Promise<TileDoc> {
+    return this._ceramic.loadStream<TileDoc>(id)
   }
 
   async _createRecord(
     definition: DefinitionWithID,
-    content: unknown,
+    content: Record<string, any>,
     { pin }: CreateOptions = {}
-  ): Promise<DocID> {
-    const metadata = { controllers: [this.id], family: definition.id.toString() }
+  ): Promise<StreamID> {
+    const metadata = {
+      deterministic: true,
+      controllers: [this.id],
+      family: definition.id.toString(),
+    }
     // Doc must first be created in a deterministic way
-    const doc = await this._ceramic.createDocument(
-      'tile',
-      { deterministic: true, metadata },
-      { anchor: false, publish: false }
-    )
+    const doc = await TileDocument.create<TileContent>(this._ceramic, null, metadata, {
+      anchor: false,
+      publish: false,
+    })
     // Then be updated with content and schema
-    const updated = doc.change({ content, metadata: { ...metadata, schema: definition.schema } })
+    const updated = doc.update(content, { ...metadata, schema: definition.schema })
     if (pin ?? this._autopin) {
       await Promise.all([updated, this._ceramic.pin.add(doc.id)])
     } else {
@@ -289,23 +308,32 @@ export class IDX {
     return index?.[key] ?? null
   }
 
-  async _setReference(key: IndexKey, id: DocID): Promise<void> {
-    await this._indexProxy.changeContent<Index>((index) => {
+  async _setReference(key: IndexKey, id: StreamID): Promise<void> {
+    await this._indexProxy.changeContent((index) => {
       return { ...index, [key]: id.toUrl() }
     })
   }
 
   async _setReferences(references: Index): Promise<void> {
     if (Object.keys(references).length !== 0) {
-      await this._indexProxy.changeContent<Index>((index) => {
+      await this._indexProxy.changeContent((index) => {
         return { ...index, ...references }
       })
     }
   }
 
   async _removeReference(key: IndexKey): Promise<void> {
-    await this._indexProxy.changeContent<Index>(({ [key]: _remove, ...index }) => {
+    await this._indexProxy.changeContent((index) => {
+      if (index) delete index[key]
       return index
     })
+  }
+
+  async caip10ToDid(accountId: string): Promise<string> {
+    const link = await Caip10Link.fromAccount(this._ceramic, accountId)
+    if (link.did == null) {
+      throw new Error(`No DID found for ${accountId}`)
+    }
+    return link.did
   }
 }