+++ packages/core/src/types.ts
@@ -1,11 +1,10 @@
-import type DocID from '@ceramicnetwork/docid'
+import type StreamID from '@ceramicnetwork/streamid'
 import type { Definition } from '@ceramicstudio/idx-constants'
 
 export type Aliases = Record<string, string>
 
-export type DefinitionWithID<
-  C extends Record<string, unknown> = Record<string, unknown>
-> = Definition<C> & { id: DocID }
+export type DefinitionWithID<C extends Record<string, unknown> = Record<string, unknown>> =
+  Definition<C> & { id: StreamID }
 
 export type IndexKey = string
 export type Index = Record<IndexKey, string>