+++ packages/tools/src/docset.ts
@@ -1,5 +1,6 @@
-import type { CeramicApi, DocMetadata, Doctype } from '@ceramicnetwork/common'
-import type DocID from '@ceramicnetwork/docid'
+import type { CeramicApi, StreamMetadata } from '@ceramicnetwork/common'
+import type { StreamRef } from '@ceramicnetwork/streamid'
+import type { TileDocument } from '@ceramicnetwork/stream-tile'
 import type { Field, GraphQLDocSetRecords } from '@ceramicstudio/idx-graphql-types'
 import { camelCase, pascalCase } from 'change-case'
 import type { DagJWSResult } from 'dids'
@@ -14,8 +15,8 @@ type CeramicMeta = { type: 'tile'; schema?: string | Array<string> }
 type CeramicSchema = Schema & { $ceramic?: CeramicMeta }
 
 export type CreatedDoc = {
-  id: DocID
-  dependencies: Array<DocID>
+  id: StreamRef
+  dependencies: Array<StreamRef>
 }
 
 export type PublishedDocSet = {
@@ -125,7 +126,7 @@ function extractSchemaReferences(schema: Schema): Array<string> {
 
 export class DocSet {
   _ceramic: CeramicApi
-  _docs: Record<string, Promise<Doctype>> = {}
+  _docs: Record<string, Promise<TileDocument>> = {}
   _definitions: Record<string, Promise<CreatedDoc>> = {}
   _schemas: Record<string, Promise<CreatedDoc>> = {}
   _schemaAliases: Record<string, string> = {}
@@ -138,14 +139,14 @@ export class DocSet {
     this._ceramic = ceramic
   }
 
-  async loadCreated(created: Promise<CreatedDoc>): Promise<Doctype> {
+  async loadCreated(created: Promise<CreatedDoc>): Promise<TileDocument> {
     return await this.loadDoc((await created).id)
   }
 
-  async loadDoc(docID: DocID | string): Promise<Doctype> {
-    const id = docIDToString(docID)
+  async loadDoc(streamID: StreamRef | string): Promise<TileDocument> {
+    const id = docIDToString(streamID)
     if (this._docs[id] == null) {
-      this._docs[id] = this._ceramic.loadDocument(id)
+      this._docs[id] = this._ceramic.loadStream<TileDocument>(id)
     }
     return await this._docs[id]
   }
@@ -181,7 +182,7 @@ export class DocSet {
   createSchema(
     name: string,
     schema: Schema,
-    deps: Array<Promise<DocID>> = []
+    deps: Array<Promise<StreamRef>> = []
   ): Promise<CreatedDoc> {
     if (this.hasSchema(name)) {
       throw new Error(`Schema ${name} already exists`)
@@ -202,7 +203,7 @@ export class DocSet {
     return this._schemas[name]
   }
 
-  async addSchema(schema: Schema, alias?: string): Promise<DocID> {
+  async addSchema(schema: Schema, alias?: string): Promise<StreamRef> {
     const name = alias ?? (schema.title as string | undefined)
     if (name == null) {
       throw new Error('Schema must have a title property or an alias must be provided')
@@ -217,7 +218,7 @@ export class DocSet {
     return created.id
   }
 
-  async useExistingSchema(id: DocID | string, alias?: string): Promise<DocID> {
+  async useExistingSchema(id: StreamRef | string, alias?: string): Promise<StreamRef> {
     const existingAlias = this._schemaAliases[docIDToString(id)]
     if (existingAlias != null) {
       const existing = this._schemas[existingAlias]
@@ -270,7 +271,7 @@ export class DocSet {
   createDefinition(
     alias: string,
     definition: Definition,
-    deps: Array<Promise<DocID>> = []
+    deps: Array<Promise<StreamRef>> = []
   ): Promise<CreatedDoc> {
     if (this.hasDefinition(alias)) {
       throw new Error(`Definition ${alias} already exists`)
@@ -288,7 +289,7 @@ export class DocSet {
     return this._definitions[alias]
   }
 
-  async addDefinition(definition: Definition, alias = definition.name): Promise<DocID> {
+  async addDefinition(definition: Definition, alias = definition.name): Promise<StreamRef> {
     const created = await this.createDefinition(alias, definition, [
       this.useExistingSchema(definition.schema),
     ])
@@ -314,8 +315,8 @@ export class DocSet {
   createTile<T extends Record<string, unknown>>(
     alias: string,
     contents: T,
-    meta: Partial<DocMetadata>,
-    deps: Array<Promise<DocID>> = []
+    meta: Partial<StreamMetadata>,
+    deps: Array<Promise<StreamRef>> = []
   ): Promise<CreatedDoc> {
     if (this.hasTile(alias)) {
       throw new Error(`Tile ${alias} already exists`)
@@ -336,8 +337,8 @@ export class DocSet {
   async addTile<T extends Record<string, unknown>>(
     alias: string,
     contents: T,
-    meta: Partial<DocMetadata>
-  ): Promise<DocID> {
+    meta: Partial<StreamMetadata>
+  ): Promise<StreamRef> {
     if (meta.schema == null) {
       throw new Error('Missing schema to add tile')
     }
@@ -382,7 +383,8 @@ export class DocSet {
       dependencies.forEach((depid) => {
         deps.add(depid.toString())
       })
-      const commits = await this._ceramic.loadDocumentCommits(id)
+      const streamid = id.baseID
+      const commits = await this._ceramic.loadStreamCommits(streamid)
       docs[id.toString()] = commits.map((r) => r.value as DagJWSResult)
     }
 
@@ -474,8 +476,8 @@ export async function publishSignedDocSet(
   ceramic: CeramicApi,
   docSet: SignedDocSet
 ): Promise<void> {
-  const schemas: Array<Promise<Doctype>> = []
-  const others: Array<Promise<Doctype>> = []
+  const schemas: Array<Promise<TileDocument>> = []
+  const others: Array<Promise<TileDocument>> = []
 
   Object.entries(docSet.docs).forEach(([id, commits]) => {
     const publish = publishCommits(ceramic, commits)